requirements.txt
tensorflow>=2.0.0
streamlit>=1.0.0
pillow>=8.0.0
numpy>=1.19.2
pytest>=6.0.0